==> 04_analysis_word_lists_original_datasets.py <==
import matplotlib
import matplotlib.pyplot as plt

import pandas as pd
from src.evaluators.word_list import WordListEvaluator
from src.constants import COLUMN_TEXT_ID, COLUMN_TEXT, COLUMN_TEXT_LEVEL, CEFR_LEVELS
from constants import CERD, CAM_MCQ
from src.utils_plotting import line_plot_word_lists_count

matplotlib.rcParams['mathtext.fontset'] = 'stix'
matplotlib.rcParams['font.family'] = 'STIXGeneral'
matplotlib.rcParams['font.size'] = 14


def get_word_lists_per_text(df):
    texts_dict = {text_id: text for text_id, text in df[[COLUMN_TEXT_ID, COLUMN_TEXT]].values}
    target_level_dict = {text_id: level for text_id, level in df[[COLUMN_TEXT_ID, COLUMN_TEXT_LEVEL]].values}
    word_lists = WordListEvaluator().measure_word_frequency(texts_dict)
    word_lists[COLUMN_TEXT_LEVEL] = word_lists.apply(lambda r: target_level_dict[r[COLUMN_TEXT_ID]], axis=1)
    for level in CEFR_LEVELS:
        word_lists[level + '_frac'] = word_lists.apply(lambda r: r[level]/r['text_length_n_words'], axis=1)
    word_lists_per_level = [word_lists[word_lists[COLUMN_TEXT_LEVEL] == level] for level in CEFR_LEVELS]
    return word_lists_per_level


def boxplot_text_length(word_lists_per_level, title, filename=None):
    fig, ax = plt.subplots(figsize=(6, 4.2))
    ax.boxplot([local_df['text_length_n_words'] for local_df in word_lists_per_level])
    ax.set_title(f"Text length (n. words) | {title}")
    ax.set_xticks(range(1, len(CEFR_LEVELS)+1))
    ax.set_xticklabels(CEFR_LEVELS)
    ax.grid(axis='y')
    plt.tight_layout()
    if filename is None:
        plt.show()
    else:
        plt.savefig(f'output_figures/boxplot_text_length/{filename}.png')
    plt.close(fig)


def boxplot_count_words_per_level(word_lists_per_level, title, mode='frac', filename=None):
    if mode not in ['frac', 'count']:
        raise ValueError(f"Mode must be either 'frac' or 'count', not '{mode}'")
    for level in CEFR_LEVELS:
        fig, ax = plt.subplots(figsize=(6, 4.2))
        if mode == 'frac':
            ax.boxplot([local_df[f'{level}_{mode}'] for local_df in word_lists_per_level])
            ax.set_ylabel(f"Fraction of words in text.")
        else:
            ax.boxplot([local_df[f'{level}'] for local_df in word_lists_per_level])
            ax.set_ylabel(f"N. of words in text.")
        ax.set_title(f"Words from level {level} | {title}")
        ax.set_xticks(range(1, len(CEFR_LEVELS)+1))
        ax.set_xticklabels(CEFR_LEVELS)
        ax.grid(axis='y')
        plt.tight_layout()
        if filename is None:
            plt.show()
        else:
            plt.savefig(f'output_figures/boxplot_count_words_per_level/{filename}_{level}_{mode}.png')
        plt.close(fig)


def main():
    df_cerd = pd.read_csv('data/input/cerd.csv')
    df_cam_mcq = pd.read_csv('data/input/mcq_cupa.csv')

    word_lists_count_per_level_cerd = get_word_lists_per_text(df_cerd)
    word_lists_count_per_level_cam_mcq = get_word_lists_per_text(df_cam_mcq)
    word_lists_count_per_level_aggregate = get_word_lists_per_text(pd.concat([df_cerd, df_cam_mcq], axis=0))

    boxplot_text_length(word_lists_count_per_level_cerd, title='CERD', filename=CERD)
    boxplot_text_length(word_lists_count_per_level_cam_mcq, title='Cambridge MCQ', filename=CAM_MCQ)
    boxplot_text_length(word_lists_count_per_level_aggregate, title='Aggregate', filename='aggregate')

    for mode in ['frac', 'count']:
        boxplot_count_words_per_level(word_lists_count_per_level_cerd, title='CERD', mode=mode, filename=CERD)
        boxplot_count_words_per_level(word_lists_count_per_level_cam_mcq, title='Cambridge MCQ', mode=mode, filename=CAM_MCQ)
        boxplot_count_words_per_level(word_lists_count_per_level_aggregate, title='Aggregate', mode=mode, filename='aggregate')

    line_plot_word_lists_count(word_lists_count_per_level_cerd, title='CERD', filename=CERD)
    line_plot_word_lists_count(word_lists_count_per_level_cam_mcq, title='Cambridge MCQ', filename=CAM_MCQ)
    line_plot_word_lists_count(word_lists_count_per_level_aggregate, title='Aggregate', filename='aggregate')

    # This is to print the frequency of words from different levels in the reference data.
    out_df = pd.DataFrame(columns=['text_cefr', 'word_level', 'avg_frac'])
    for level in CEFR_LEVELS:  # This is the CEFR level of the words from the wordlist.
        for idx, local_df in enumerate(word_lists_count_per_level_aggregate):
            # this is for the cefr level of the reading passage (?)
            new_row = {
                'text_cefr': [CEFR_LEVELS[idx]],
                'word_level': [level],
                'avg_frac': [local_df[level + '_frac'].mean()],
            }
            if len(out_df) == 0:
                out_df = pd.DataFrame(new_row)
            else:
                out_df = pd.concat([out_df, pd.DataFrame(new_row)], ignore_index=True)
    print(out_df)


if __name__ == '__main__':
    main()
